platform_api/requirements.txt
fastapi>=0.115.0,<1
python-dotenv>=1.0.1,<2
uvicorn[standard]>=0.32.0,<1
httpx>=0.27.0,<1
google-genai>=0.3.0
PyJWT[crypto]>=2.8.0,<3
google-auth>=2.29.0,<3
passlib[bcrypt]>=1.7.4,<2
bcrypt>=4.0.0,<5
redis>=5.0.0,<6
celery>=5.4.0,<6
prometheus-client>=0.21.0,<1
